envirotrackapp/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
django-debug-toolbar==4.2.0
